refactor: moved delete_jobs_and_artifacts to job_util.py
